Add command group wrappers for aiming
